Created a repo and added the asteroids code
Committing my asteroids code.  This is very much still a work in
progress.  You can fly the ship and shoot asteroids.  Crashing into an
asteroid is not detected.  Asteroids disappear instead of breaking apart
when you shoot them.  And there is no scoring system.  Also, the code is
kind of hacked together right now, so I need to refactor.